Changed Scoring System Structure
Before, the score was stored in the Level gameObject, but it is now a field of the Player gameObject. This was done so classes do not have to talk to each other so much.

These fields have been moved from the Level class to the Player class:
- score
- cachedScore (now called previousPlatformIndex, the logic has changed so that the amount of platforms passed is calculated by the difference in indices rather than "if the player is above or below each and every platform")
- cachedBounces (now cachedBorderBounces)

Previously, the Platform gameObject was handling when to add to "CachedScore," which belonged to the Level class. Now, it, along with all other score logic, is calculated by the Player class.

Now, when level.UpdateScore() is called, it simply gets the Score property from player, and assigns it to scoreText.text.