Added Volume and FPSToggle options
In the Start Menu scene, there is now an "Options" button instead of a "Quit" button (This was a placeholder for windows users, but now mobile is focused).

The "Options" button, when pressed, shows options to control the volumes of different sources and a toggle that enables/disables the FPSText object via a new script attached to the FPSText GameObject.

There is now a "UserSettings" static class in the namespace "Game.Settings," which allows for easy addition for new user settings down the line.